Start Orbot via always-on VPN without onboarding
When the system tries to start Orbot, if the onboarding process has yet
to be completed, automatically connect to Tor without requiring user
intervention. This allows Orbot to get started automatically when it is
bundled with the OS in always-on mode.

Test: Manual: Install Orbot and connect to Tor. In the system settings,
set Orbot as an always-on VPN. Clear storage for Orbot. Reboot the
device. Orbot should automatically connect and be available for use.