Fix error with minify when using AGP 8
Adding a common -dontwarn for java.lang.invoke.StringConcatFactory, or for all
serialization classes does not solve the problem, but hide it in the serialization
itself, or in the unrelated code of other dependencies.

An alternative to using the -dontwarn in the ProGuard rules for building the library
is to turn off generation of invokedynamic for string concatenation for kotlinc when
building the with the option -Xstring-concat=inline.

AGP already turns off off generation of invokedynamic for string concatenation for javac.

Fixes:

> Task :appcore:minifyReleaseWithR8

ERROR: R8: Missing class java.lang.invoke.StringConcatFactory
(referenced from: void org.torproject.android.core.Languages.<init>(android.app.Activity)
and 1 other context)